fast_api_backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
aiofiles
python-jose[cryptography]
passlib[bcrypt]
